Refactor comments component
